# python-flask/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Ytel API version 3"
  version: "3.12"
  title: "Ytel API V3"
host: "api.ytel.com"
basePath: "/api/v3"
schemes:
- "https"
consumes:
- "application/x-www-form-urlencoded"
produces:
- "application/json"
paths:
  /template/view.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "View Template"
      description: "View a Shared ShortCode Template"
      operationId: "template_view"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "TemplateId"
        in: "formData"
        description: "The unique identifier for a template object"
        required: true
        type: "string"
        format: "uuid"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
...
  /transcriptions/viewtranscription.json:
    post:
      tags:
      - "Transcription"
      summary: "View Transcription"
      description: "Retrieve information about a transaction by its TranscriptionSid."
      operationId: "transcriptions_viewtranscription"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "transcriptionsid"
        in: "formData"
        description: "The unique identifier for a transcription object."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.transcription_controller"
  /transcriptions/recordingtranscription.json:
    post:
      tags:
      - "Transcription"
      summary: "Transcribe Recording"
      description: "Transcribe a recording by its RecordingSid."
      operationId: "transcriptions_recordingtranscription"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "recordingSid"
        in: "formData"
        description: "The unique identifier for a recording object."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.transcription_controller"
  /transcriptions/audiourltranscription.json:
    post:
      tags:
      - "Transcription"
      summary: "Transcribe Audio URL"
      description: "Transcribe an audio recording from a file."
      operationId: "transcriptions_audiourltranscription"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "audiourl"
        in: "formData"
        description: "URL pointing to the location of the audio file that is to be\
          \ transcribed."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.transcription_controller"
...
  /carrier/lookuplist.json:
    post:
      tags:
      - "Carrier"
      summary: "Carrier Results"
      description: "Retrieve a list of carrier lookup objects."
      operationId: "carrier_lookuplist"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "Page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "PageSize"
        in: "formData"
        description: "The count of objects to return per page."
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.carrier_controller"
...
  /carrier/lookup.json:
    post:
      tags:
      - "Carrier"
      summary: "Lookup Carrier"
      description: "Get the Carrier Lookup"
      operationId: "carrier_lookup"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "PhoneNumber"
        in: "formData"
        description: "A valid 10-digit number (E.164 format)."
        required: true
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.carrier_controller"
...
  /transcriptions/listtranscription.json:
    post:
      tags:
      - "Transcription"
      summary: "List Transcriptions"
      description: "Retrieve a list of transcription objects for your Ytel account."
      operationId: "transcriptions_listtranscription"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "The count of objects to return per page."
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "status"
        in: "formData"
        description: "The state of the transcription."
        required: false
        type: "string"
        enum:
        - "inprogress"
        - "success"
        - "failure"
        x-enum-elements:
        - name: "INPROGRESS"
          description: ""
        - name: "Success"
          description: ""
        - name: "Failure"
          description: ""
      - name: "dateTranscribed"
        in: "formData"
        description: "The date the transcription took place."
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.transcription_controller"
...
  /template/lists.json:
    post:
      tags:
      - "SharedShortCode"
      summary: "List Templates"
      description: "List Shortcode Templates by Type"
      operationId: "template_lists"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "type"
        in: "formData"
        description: "The type (category) of template Valid values: marketing, authorization"
        required: false
        type: "string"
        default: "authorization"
      - name: "page"
        in: "formData"
        description: "The page count to retrieve from the total results in the collection.\
          \ Page indexing starts at 1."
        required: false
        type: "integer"
        default: 1
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "pagesize"
        in: "formData"
        description: "The count of objects to return per page."
        required: false
        type: "integer"
        default: 10
        exclusiveMaximum: false
        exclusiveMinimum: false
        format: "int32"
      - name: "Shortcode"
        in: "formData"
        description: "Only list templates of type"
        required: false
        type: "string"
      responses:
        200:
          description: ""
          examples: {}
          schema:
            type: "string"
      security:
      - auth: []
      x-unitTests: []
      x-operation-settings:
        CollectParameters: true
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-swagger-router-controller: "swagger_server.controllers.shared_short_code_controller"
...
securityDefinitions:
  auth:
    type: "basic"
    x-testValue: []
    x-skip-client-authentication: false
...